requirements.txt
google-generativeai>=0.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0

